requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-dotenv
google-generativeai
watchdog